requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.5.0
dill==0.3.8
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
intel_extension_for_pytorch==2.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oneccl-bind-pt==2.6.0+cpu
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tqdm==4.67.1
transformers==4.51.1
triton==3.2.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.19.0
